multiplication can be performmed with a continuous function and an AudioFile without losing precision until the new AudioFile is created

+++ imaudio/func/AdsrEnvelope.cpp
@@ -9,7 +9,7 @@ AdsrEnvelope::AdsrEnvelope(double Attack, double Decay, double Sustain, double R
 	length = Length;
 }
 
-long AdsrEnvelope::sampleAtTime(double time) {
+double AdsrEnvelope::sampleAtTime(double time) {
 	if (time < 0.0) {
 		return 0.0;
 	}
