+++ imaudio/func/TriangleWave.h
@@ -7,7 +7,7 @@ class TriangleWave : public iWaveform {
 public:
 	TriangleWave(double Amplitude, double Frequency) : iWaveform(Amplitude, Frequency) { }
 
-	virtual long sampleAtTime(double time);
+	virtual double sampleAtTime(double time);
 
 	inline virtual string functionName() {
 		return "trianglewave";
